Rename folder name task to tasks
