+++ src/core/tasks/tasks-switch-element.ts
